SB-631: Move the misplaced configuration classes under org.carlspring.strongbox.config
Implemented!
