+++ strongbox-storage/strongbox-storage-api/src/test/java/org/carlspring/strongbox/BaseStorageApiTest.java
@@ -1,6 +1,7 @@
 package org.carlspring.strongbox;
 
 import org.carlspring.strongbox.config.ClientConfig;
+import org.carlspring.strongbox.config.CommonConfig;
 import org.carlspring.strongbox.config.StorageApiConfig;
 import org.carlspring.strongbox.testing.TestCaseWithArtifactGeneration;
 import org.springframework.context.annotation.Import;
