+++ strongbox-storage/strongbox-storage-api/src/test/java/org/carlspring/strongbox/storage/validation/version/VersionValidatorServiceTest.java
@@ -1,6 +1,6 @@
 package org.carlspring.strongbox.storage.validation.version;
 
-import org.carlspring.strongbox.CommonConfig;
+import org.carlspring.strongbox.config.CommonConfig;
 import org.carlspring.strongbox.config.StorageApiConfig;
 import org.carlspring.strongbox.services.VersionValidatorService;
