--- strongbox-testing/strongbox-testing-core/src/main/java/org/carlspring/strongbox/TestingCoreConfig.java
@@ -1,21 +1,0 @@
-package org.carlspring.strongbox;
-
-import org.carlspring.strongbox.testing.AssignedPorts;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * TestingCoreConfig.
- */
-@Configuration
-public class TestingCoreConfig
-{
-
-    @Bean(name = "assignedPorts")
-    AssignedPorts assignedPorts()
-    {
-        return new AssignedPorts();
-    }
-
-}
