+++ strongbox-user-management/src/test/java/org/carlspring/strongbox/users/UserServiceTestContext.java
@@ -1,6 +1,7 @@
 package org.carlspring.strongbox.users;
 
 import org.carlspring.strongbox.config.DataServiceConfig;
+import org.carlspring.strongbox.config.UsersConfig;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
