+++ strongbox-web-core/src/main/java/org/carlspring/strongbox/config/WebConfig.java
@@ -1,9 +1,7 @@
 package org.carlspring.strongbox.config;
 
-import org.carlspring.strongbox.CommonConfig;
 import org.carlspring.strongbox.StorageIndexingConfig;
 import org.carlspring.strongbox.configuration.StrongboxSecurityConfig;
-import org.carlspring.strongbox.users.UsersConfig;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
